+++ CHANGELOG.md
@@ -63,6 +63,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: `Agclos_t.strdict` is now an opaque pointer.
 - **Breaking**: `agattr` can no longer be used to create attributes with
   HTML-like values. For this, use `agattr_html` instead.
+- **Breaking**: `agstrfree` takes an extra parameter indicating whether the
+  string being freed is an HTML-like string.
 
 ### Fixed